company servislerinde ki int degerler longa cevrildi.

+++ maindata-service/src/main/java/com/oee/repository/CompanyInfoRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.oee.entity.CompanyInfo;
 
-public interface CompanyInfoRepository extends JpaRepository<CompanyInfo, Integer>{
+public interface CompanyInfoRepository extends JpaRepository<CompanyInfo, Long>{
 	
 }
